Finish hangman game
The code was divided into a class with two modules. The class Hangman is in charge of all functions related to checking if a letter guess by the user is inside the hidden word of the computer. This class provides the option to save the game. The module LoadGame, loads the information of a previous game and starts the game. The module StartOrLoadGame is called at the beginning of the game. Checks if there are games available to load or starts a new game